Feature spec for admin adding notes to participant
